Derive containing-block claims from CoreStyle::is_containing_block
Taffy's claim policy moved from a LayoutContainingBlock::oof_claims tree
hook to CoreStyle::is_containing_block (DioxusLabs/taffy#1170/#1172), so
that the grid algorithm's static-position emission and the positioning
pass share one source of truth. Implement it on TaffyStyloStyle, moving
the style-only fixed/absolute containing-block predicates into
stylo_taffy::convert, and reuse the same claims in the damage pass.

Bumps taffy to the rewritten stack top (5e71ce15).
